ast: Make JsonEqual a test helper only
This is no longer needed in cmd packages.
