+++ v1/util/jsonv2.go
@@ -7,26 +7,9 @@
 package util
 
 import (
-	"bytes"
 	"encoding/json/jsontext"
 	"encoding/json/v2"
 )
-
-type stringOrBytes interface {
-	string | []byte
-}
-
-// JsonEqual returns true if the canonical JSON encoding of a and b are equal,
-// meaning that a and b are compared without regard to things like whitespace,
-// key order, etc. For more details, see [jsontext.Value.Canonicalize].
-func JsonEqual[A, B stringOrBytes](a A, b B) bool {
-	v1, v2 := jsontext.Value(a), jsontext.Value(b)
-
-	v1.Canonicalize()
-	v2.Canonicalize()
-
-	return bytes.Equal(v1, v2)
-}
 
 // WriteMarshalerToArray writes the JSON array of items to the encoder.
 func WriteMarshalerToArray[T json.MarshalerTo](e *jsontext.Encoder, items []T) error {
